refactor(doctor): extract mutable allowlist scanning
